+++ src/functionalTest/groovy/com/autonomousapps/android/ServiceLoaderSpec.groovy
@@ -1,8 +1,6 @@
 package com.autonomousapps.android
 
 import com.autonomousapps.fixtures.ServiceLoaderProject
-import com.autonomousapps.internal.android.AgpVersion
-import spock.lang.Ignore
 import spock.lang.Unroll
 
 import static com.autonomousapps.utils.Runner.build
@@ -11,7 +9,6 @@ import static com.google.common.truth.Truth.assertThat
 @SuppressWarnings("GroovyAssignabilityCheck")
 final class ServiceLoaderSpec extends AbstractAndroidSpec {
 
-  @Ignore("This test will fail until the work on service loaders is complete")
   @Unroll
   def "service-loading libraries are not reported as unused (#gradleVersion AGP #agpVersion)"() {
     given:
@@ -27,6 +24,6 @@ final class ServiceLoaderSpec extends AbstractAndroidSpec {
     assertThat(expectedAdviceForApp).containsExactlyElementsIn(actualAdviceForApp)
 
     where:
-    [gradleVersion, agpVersion] << gradleAgpMatrix(AgpVersion.version('3.6'))
+    [gradleVersion, agpVersion] << gradleAgpMatrix()
   }
 }